Tweak scheduler score
Now the time since last activity is "positive", i.e. the bigger the time,
the greater chance of selecting this backend.

The point of this change is to keep scheduler scores roughly decreasing in
round-robin order, and to keep the iteration count of the main loop of
ngx_http_upstream_choose_fair_peer as short as possible.